internet_shop/int_shop/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
bleach==6.0.0
boto3==1.27.0
botocore==1.30.0
cachetools==5.3.0
captcha==0.5.0
celery==5.2.7
celery-amqp-backend==1.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
coverage==7.2.7
cryptography==40.0.2
defusedxml==0.7.1
Django==4.1.7
django-bootstrap5==23.1
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-extensions==3.2.1
django-filter==23.2
django-storages==1.13.2
django-summernote==0.8.20.0
djangorestframework==3.14.0
drf-yasg==1.21.7
flake8==6.0.0
flower==1.2.0
hiredis==2.2.2
httplib2==0.22.0
humanize==4.6.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kombu==5.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
phonenumbers==8.13.11
Pillow==10.0.1
pluggy==1.2.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.3
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0.1
redis==4.5.4
referencing==0.30.2
requests==2.29.0
requests-oauthlib==1.3.1
rpds-py==0.10.3
rsa==4.9
s3transfer==0.6.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.3
stripe==5.4.0
tornado==6.2
uritemplate==4.1.1
urllib3==1.26.14
uWSGI==2.0.21
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.2
